// src/types/openapi/generated.d.ts
export interface paths {
    "/broadcasters": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get broadcasters
         * @description Get information from the broadcasters table.
         */
        get: {
            parameters: {
                query: {
                    /**
                     * @description The wikis you want data from. Pipe-separate multiple wikis for multiwiki requests.
                     *
                     *     **Example:** `dota2`, `dota2|counterstrike`
                     */
                    wiki: components["parameters"]["wiki"];
                    /**
                     * @description The filters you want to apply to the request.
                     *
                     *     **Example:** `[[pagename::Some/Liquipedia/Page]] AND [[namespace::0]]`
                     */
                    conditions?: components["parameters"]["conditions"];
                    /**
                     * @description The datapoints you want to query.
                     *
                     *     **Example:** `pagename, pageid, namespace`
                     */
                    query?: components["parameters"]["query"];
                    /**
                     * @description The amount of results you want.
                     *
                     *     **Example:** `20`
                     */
                    limit?: components["parameters"]["limit"];
                    /**
                     * @description This can be used for pagination.
                     *
                     *     **Example:** `20`
                     */
                    offset?: components["parameters"]["offset"];
                    /**
                     * @description The order you want your result in.
                     *
                     *     **Example:** `pagename ASC`
                     */
                    order?: components["parameters"]["order"];
                    /**
                     * @description What you want your results grouped by (this can be helpful when using aggregate functions).
                     *
                     *     **Example:** `pagename ASC`
                     */
                    groupby?: components["parameters"]["groupby"];
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful call */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["successfulResponse"] | components["schemas"]["successfulResponseWithWarning"];
                    };
                };
                /** @description Invalid API key */
                403: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["error"];
                    };
                };
                /** @description Asking for data that does not exist */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["error"];
                    };
                };
                /** @description Over API limit */
                429: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["error"];
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /** Successful call */
        successfulResponse: {
            result?: Record<string, never>[];
        };
        /** Successful call with warning */
        successfulResponseWithWarning: {
            result?: Record<string, never>[];
            warning?: string[];
        };
        /** Error */
        error: {
            error?: string[];
        };
    };
    responses: never;
    parameters: {
        /**
         * @description The wikis you want data from. Pipe-separate multiple wikis for multiwiki requests.
         *
         *     **Example:** `dota2`, `dota2|counterstrike`
         */
        wiki: string;
        /**
         * @description The filters you want to apply to the request.
         *
         *     **Example:** `[[pagename::Some/Liquipedia/Page]] AND [[namespace::0]]`
         */
        conditions: string;
        /**
         * @description The datapoints you want to query.
         *
         *     **Example:** `pagename, pageid, namespace`
         */
        query: string;
        /**
         * @description The amount of results you want.
         *
         *     **Example:** `20`
         */
        limit: number;
        /**
         * @description This can be used for pagination.
         *
         *     **Example:** `20`
         */
        offset: number;
        /**
         * @description The order you want your result in.
         *
         *     **Example:** `pagename ASC`
         */
        order: string;
        /**
         * @description What you want your results grouped by (this can be helpful when using aggregate functions).
         *
         *     **Example:** `pagename ASC`
         */
        groupby: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}